Refactor: dto 패키지를 presentation에서 분리
