+++ src/main/java/com/toty/email/application/EmailService.java
@@ -2,11 +2,10 @@ package com.toty.email.application;
 
 import com.toty.email.domain.EmailMessage;
 import com.toty.user.application.UserService;
-import com.toty.user.domain.UserRepository;
+import com.toty.user.domain.repository.UserRepository;
 import java.util.Random;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
 import org.thymeleaf.context.Context;
 import org.thymeleaf.spring6.SpringTemplateEngine;
