+++ src/main/java/com/toty/user/domain/model/Site.java
@@ -1,4 +1,4 @@
-package com.toty.user.domain;
+package com.toty.user.domain.model;
 
 public enum Site {
     Github, Blog
