+++ src/main/java/com/toty/user/presentation/UserApiController.java
@@ -2,13 +2,13 @@ package com.toty.user.presentation;
 
 import com.toty.annotation.CurrentUser;
 import com.toty.following.application.FollowService;
-import com.toty.following.presentation.dto.response.FollowingListResponse;
+import com.toty.following.dto.response.FollowingListResponse;
 import com.toty.user.application.UserService;
 import com.toty.user.application.UserSignUpService;
-import com.toty.user.domain.User;
-import com.toty.user.application.dto.request.UserInfoUpdateRequest;
-import com.toty.user.application.dto.request.UserSignUpRequest;
-import com.toty.user.application.dto.response.UserInfoResponse;
+import com.toty.user.domain.model.User;
+import com.toty.user.dto.request.UserInfoUpdateRequest;
+import com.toty.user.dto.request.UserSignUpRequest;
+import com.toty.user.dto.response.UserInfoResponse;
 import jakarta.websocket.server.PathParam;
 import java.util.HashMap;
 import java.util.Map;
@@ -52,10 +52,10 @@ public class UserApiController {
     }
 
     // 나의/상대방의 정보 보기
+        // 본인인지 아닌지 확인 -> 아니면 약식 정보만 전달
     @GetMapping("/{id}")
     public ResponseEntity<UserInfoResponse> getUserInfo(@CurrentUser User user,
                                                         @PathVariable("id") Long id) {
-        // 본인인지 아닌지 확인 -> 아니면 약식 정보만 전달
         UserInfoResponse userInfo = userService.getUserInfo(user, id);
         return ResponseEntity.ok(userInfo);
     }
@@ -67,26 +67,5 @@ public class UserApiController {
                                                      @RequestPart MultipartFile imgFile) {
         userService.updateUserInfo(user.getId(), newInfo, imgFile);
         return ResponseEntity.ok("Done");
-    }
-
-    @GetMapping("{uid}/posts")
-    public ResponseEntity getPostByCategory(@PathVariable("uid") Long id) {
-        // todo
-        Map<String, String> response = new HashMap<>();
-        return new ResponseEntity<>(response, HttpStatus.CREATED);
-    }
-
-    // 팔로워 목록 조회
-    @GetMapping("/{uid}/followers")
-    public ResponseEntity<FollowingListResponse> followersList(@PathParam("uid") Long id, @RequestParam(value = "p", defaultValue = "1") int page) {
-        FollowingListResponse response = followService.pagedFollowings(id, true, page);
-        return ResponseEntity.ok(response);
-    }
-
-    // 팔로잉 목록 조회
-    @GetMapping("/{uid}/followings")
-    public ResponseEntity<FollowingListResponse> followingList(@PathParam("uid") Long id, @RequestParam(value = "p", defaultValue = "1") int page) {
-        FollowingListResponse response = followService.pagedFollowings(id, false, page);
-        return ResponseEntity.ok(response);
     }
 }
