+++ src/main/java/com/toty/user/presentation/UserViewController.java
@@ -2,8 +2,8 @@ package com.toty.user.presentation;
 
 import com.toty.annotation.CurrentUser;
 import com.toty.user.application.UserService;
-import com.toty.user.domain.User;
-import com.toty.user.application.dto.response.UserInfoResponse;
+import com.toty.user.domain.model.User;
+import com.toty.user.dto.response.UserInfoResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
