+++ src/test_apiE2E/java/UserApiE2ETest.java
@@ -1,6 +1,6 @@
-import com.toty.user.domain.User;
-import com.toty.user.domain.UserRepository;
-import com.toty.user.application.dto.request.UserSignUpRequest;
+import com.toty.user.domain.model.User;
+import com.toty.user.domain.repository.UserRepository;
+import com.toty.user.dto.request.UserSignUpRequest;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.hamcrest.Matchers;
