+++ src/test_unit/java/service/UserServiceTest.java
@@ -1,8 +1,8 @@
 //package service;
 //
 //import com.toty.user.application.UserService;
-//import com.toty.user.domain.User;
-//import com.toty.user.domain.UserRepository;
+//import com.toty.user.domain.model.User;
+//import com.toty.user.domain.repository.UserRepository;
 //import com.toty.user.presentation.dto.request.UserSignUpRequest;
 //import com.toty.user.presentation.dto.response.UserInfoResponse;
 //import org.assertj.core.api.Assertions;
